requirements.txt
ultralytics==8.3.92
torch 
pandas 
openpyxl
scikit-learn
exifread
mlflow
folium
exif
defusedxml
PyYAML==6.0.2
tqdm==4.66.1
azure-storage-blob==12.19.0